sync examples for orchestartion refactoring
